CONTROL ON NUMBER OF ADDS TO SONG BANK ADDED
I WANT TO CRY THIS MAKES ME SO HAPPY.

So basically before if you would press the « add to playlist » (well
technically its « click this » but w/e) multiple times it would add
multiple times…. but now its controlled so that you can only add it
once.  Need to add a message that pops up saying its already added to
the playlist.